organize transaction files
