+++ src/concepts/Transaction/views/TransactionBox/TransactionBox.module.scss
