+++ src/concepts/Transaction/views/home/HomePage.tsx
@@ -1,0 +1,13 @@
+import Head from "next/head";
+import { TransactionsPage } from "../TransactionsPage";
+
+export const HomePage = () => {
+  return (
+    <>
+      <Head>
+        <title>Co-Finance - Home</title>
+      </Head>
+      <TransactionsPage />
+    </>
+  );
+};
