--- src/concepts/infra/http/createTransactionRequest.ts
@@ -1,8 +1,0 @@
-import { NewTransaction } from "@/concepts/transactions/domain/Transaction";
-import { api } from "@/lib/http/api";
-
-export const createTransactionRequest = async (transaction: NewTransaction) => {
-  return api.post("/transactions", {
-    transaction: { ...transaction, price: Number(transaction.price)}
-  });
-}
