--- src/concepts/pages/home/HomePage.tsx
@@ -1,13 +1,0 @@
-import Head from "next/head";
-import { TransactionsPage } from "../transactions/TransactionsPage";
-
-export const HomePage = () => {
-  return (
-    <>
-      <Head>
-        <title>Co-Finance - Home</title>
-      </Head>
-      <TransactionsPage />
-    </>
-  );
-};
