--- src/concepts/transactions/index.ts
@@ -1,2 +1,0 @@
-export { TransactionProvider } from "./view/components/TransactionProvider";
-export * from "./view/hooks/useTransaction";
