--- src/concepts/transactions/view/hooks/useCreateTransaction.ts
@@ -1,30 +1,0 @@
-import { createTransactionRequest } from "@/concepts/infra/http/createTransactionRequest";
-import { useBudget } from "../../../Budget/hooks/useBudget";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { NewTransaction } from "../../domain/Transaction";
-import { useTransaction } from "./useTransaction";
-import { formatDate } from "@/concepts/utils/date";
-import { toastError } from "@/concepts/utils/toastify";
-
-export const useCreateTransaction = ({onSuccess}: any) => {
-  const { selectedMonthDate } = useTransaction()
-  const { selectedBudgetId } = useBudget()
-  const queryClient = useQueryClient()
-
-  const { mutate, isPending } = useMutation({
-    mutationFn: (newTransaction: NewTransaction) => createTransactionRequest({...newTransaction, budgetId: selectedBudgetId}),
-    onSuccess: () => {
-      onSuccess()
-      queryClient.invalidateQueries({ queryKey: ['transactions', selectedBudgetId, formatDate(selectedMonthDate)] })
-    },
-    onError: (error) => {
-      const errorMsg = error?.response?.data?.error?.details.join('. ') || 'Erro ao criar orçamento'
-      toastError(errorMsg)
-    }
-  })
-
-  return {
-    createTransaction: mutate,
-    isLoading: isPending,
-  }
-}
